auth added for goals

+++ server/controller/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../model/goalsModel')
+const User = require('../model/userModel')
 const mongoose = require('mongoose')
 
 // @desc Get Goals
 // @route GET /api/goals
 
 const getGoals = asyncHandler(async(req,res)=>{
-    const goals = await Goal.find()
+    const goals = await Goal.find({user:req.user._id})
     res.status(200).json(goals)
 })
 
@@ -20,7 +21,10 @@ const setGoals = asyncHandler(async(req,res)=>{
         throw new Error('Enter Text')
     }
 
-    const goal = await Goal.create(req.body)
+    const goal = await Goal.create({
+        text:req.body.text,
+        user:req.user._id
+    })
     res.status(201).json(goal)
 })
 
@@ -34,6 +38,19 @@ const updateGoals = asyncHandler(async(req,res)=>{
         res.status(400)
         throw new Error('No Goal Found')
     }
+
+    const user = await User.findById(req.user._id)
+
+    if(!user){
+        res.status(400)
+        throw new Error('No User Found')
+    }
+
+    if(goal.user.toString() !== user._id){
+        res.status(400)
+        throw new Error('Not Authorized')
+    }
+
     const updated = await Goal.findByIdAndUpdate(id,req.body,{new:true})
     res.status(200).json(updated)
 })
@@ -48,6 +65,19 @@ const deleteGoals = asyncHandler(async(req,res)=>{
         res.status(400)
         throw new Error('No Goal Found')
     }
+
+    const user = await User.findById(req.user._id)
+
+    if(!user){
+        res.status(400)
+        throw new Error('No User Found')
+    }
+
+    if(goal.user.toString() != user._id){
+        res.status(400)
+        throw new Error('Not Authorized')
+    }
+
     const updated = await Goal.findByIdAndDelete(id)
     res.status(200).json(updated)
 })
